requirements.txt
fastapi==0.116.1
httpx==0.28.1
langchain-core==0.3.79
langchain-ollama==0.3.10
langgraph==0.6.8
PyJWT==2.10.1
python-dotenv==1.1.1
uvicorn==0.38.0
